Moved get_response to chatgpt so it can be reused
